python/requirements.txt
sounddevice 
librosa 
matplotlib 
numpy
mplcursors
soundfile
pretty_midi
midi2audio
elevenlabs
python-dotenv
flask
noisereduce
torchcodec
demucs

# brew install fluidsynth   # macOS only

